可読性の向上
